Refactor: Consolidate results and grades pages
