+++ app/student/courses/[courseId]/grades/page.tsx
@@ -1,165 +1,156 @@
 "use client"
 
-import { GraduationCap, Eye, Download } from "lucide-react"
-import Link from "next/link"
-import { Button } from "@/components/ui/button"
-import { useEffect, useState } from "react"
-import { fetchStudentGrades } from "@/lib/api/student"
-
-interface GradeRow {
-  _id: string
-  title?: string
-  score?: number
-  max_points?: number
-  status?: string
-  due_date?: string
-  submitted_at?: string
-}
+import { useEffect, useMemo, useState } from 'react'
+import { Breadcrumb, BreadcrumbItem, BreadcrumbLink, BreadcrumbList, BreadcrumbPage, BreadcrumbSeparator } from '@/components/ui/breadcrumb'
+import ResultsTable, { ResultRow } from '@/components/assessments/ResultsTable'
+import { Skeleton } from '@/components/ui/skeleton'
+import { getQuizAttempts, listExams, getOwnExamSubmission, getQuizById, getQuizQuestions } from '@/app/lib/api'
+import { fetchModulesByCourseId } from '@/lib/api/courses'
 
 export default function CourseGradesPage({ params }: { params: { courseId: string } }) {
   const { courseId } = params
-  const [gradesData, setGradesData] = useState<GradeRow[]>([])
+  const [rows, setRows] = useState<ResultRow[]>([])
+  const [loading, setLoading] = useState(true)
 
   useEffect(() => {
-    const load = async () => {
-      try {
-        const data = await fetchStudentGrades(courseId)
-        setGradesData(Array.isArray(data) ? data : [])
-      } catch {
-        setGradesData([])
+    let mounted = true
+    async function load() {
+      // Fetch course modules to scope quiz attempts to this course
+      const [modules, quizAttemptsRes, examsRes] = await Promise.all([
+        fetchModulesByCourseId(courseId),
+        getQuizAttempts(),
+        listExams({ course: courseId as any }),
+      ])
+
+      const moduleIdSet = new Set<string>((modules as any[]).map((m) => m._id))
+
+      const out: ResultRow[] = []
+
+      if (quizAttemptsRes.ok) {
+        // enrich with quiz title and max points from questions, scoped to course's modules
+        const enriched = await Promise.all(
+          quizAttemptsRes.data.map(async (a) => {
+            const quizId = typeof a.quiz_id === 'object'
+              // @ts-expect-error possible populated object
+              ? String(a.quiz_id._id ?? a.quiz_id)
+              : a.quiz_id
+
+            const [q, qs] = await Promise.all([
+              getQuizById(quizId as any),
+              getQuizQuestions(quizId as any),
+            ])
+
+            // Skip if quiz's module is not part of this course
+            const quizModuleId = q.ok ? (q.data as any)?.module_id : undefined
+            const moduleIdStr = typeof quizModuleId === 'object' ? String((quizModuleId as any)?._id) : String(quizModuleId || '')
+            if (!moduleIdStr || !moduleIdSet.has(moduleIdStr)) return null
+
+            let title: string
+            if (q.ok && (q.data as any)?.title) {
+              title = (q.data as any).title
+            } else if (typeof a.quiz_id === 'object') {
+              // @ts-expect-error error
+              title = a.quiz_id.title ?? String(a.quiz_id._id ?? a.quiz_id)
+            } else {
+              title = String(a.quiz_id)
+            }
+
+            const max = qs.ok
+              ? qs.data.reduce((acc, q) => acc + (q.points || 0), 0)
+              : 100
+            const score = typeof a.score === 'number' ? a.score : 0
+
+            return {
+              id: a._id,
+              type: 'quiz' as const,
+              title,
+              score,
+              maxScore: max,
+              percentage: max ? (score / max) * 100 : 0,
+              passed: max
+                ? (score / max) * 100 >= (q.ok ? (q.data as any).pass_percentage : 50)
+                : score >= 50,
+              date: (a as any).completed_at || (a as any).started_at,
+            } as ResultRow
+          })
+        )
+        out.push(...enriched.filter(Boolean) as ResultRow[])
+      }
+
+      if (examsRes.ok) {
+        const exams = examsRes.data.exams
+        await Promise.all(
+          exams.map(async (e: any) => {
+            const s = await getOwnExamSubmission(e._id)
+            if (s.ok && s.data) {
+              const total = Number(
+                s.data.totalScore ??
+                  (Number(s.data.autoScore || 0) + Number(s.data.manualScore || 0))
+              )
+              const max = Number(e.totalPoints ?? 100)
+              out.push({
+                id: s.data._id,
+                type: 'exam',
+                title: e.title,
+                score: total,
+                maxScore: max,
+                percentage: max ? (total / max) * 100 : 0,
+                passed:
+                  e.passingScore != null ? total >= e.passingScore : total >= 0.5 * max,
+                date: undefined,
+              })
+            }
+          })
+        )
+      }
+
+      if (mounted) {
+        setRows(out)
+        setLoading(false)
       }
     }
+    setLoading(true)
     load()
+    return () => {
+      mounted = false
+    }
   }, [courseId])
 
-  const calculateOverallGrade = () => {
-    const gradedAssignments = gradesData.filter(grade => grade.status === "Graded" && typeof grade.score === 'number' && typeof grade.max_points === 'number')
-    if (gradedAssignments.length === 0) return "N/A"
-    const totalPoints = gradedAssignments.reduce((sum, grade) => {
-      return sum + (grade.score as number)
-    }, 0)
-    const maxPoints = gradedAssignments.reduce((sum, grade) => {
-      return sum + (grade.max_points as number)
-    }, 0)
-    return `${totalPoints}/${maxPoints} (${Math.round((totalPoints / maxPoints) * 100)}%)`
-  }
+  const sorted = useMemo(
+    () =>
+      rows
+        .slice()
+        .sort(
+          (a, b) =>
+            new Date((b as any).date || 0).getTime() - new Date((a as any).date || 0).getTime()
+        ),
+    [rows]
+  )
 
   return (
-    <div className="flex flex-1 flex-col">
-      <header className="flex h-16 shrink-0 items-center gap-4 border-b bg-white px-4 md:px-6">
-        <h1 className="text-sm font-semibold text-gray-900">
-          <Link href={`/student/courses/${courseId}/home`} className="text-blue-600 hover:underline">
-            Communicating_for_Impact
-          </Link>{" "}
-          <span className="text-gray-400">›</span> Grades
-        </h1>
-        <div className="ml-auto flex items-center gap-2">
-          <Button variant="outline" className="h-8 text-sm bg-transparent border-gray-300">
-            <Download className="h-4 w-4 mr-2" />
-            Export Grades
-          </Button>
-          <Button variant="outline" className="h-8 text-sm bg-transparent border-gray-300">
-            <Eye className="h-4 w-4 mr-2" />
-            View All Grades
-          </Button>
-        </div>
-      </header>
-
-      <main className="flex flex-1 flex-col gap-4 p-4 md:gap-6 md:p-6">
-        {/* Overall Grade Summary */}
-        <div className="rounded-md border border-gray-200 bg-white shadow-sm p-6">
-          <div className="flex items-center gap-3 mb-4">
-            <GraduationCap className="h-6 w-6 text-blue-600" />
-            <h2 className="text-lg font-semibold text-gray-900">Grade Summary</h2>
+    <div className="p-4 md:p-6">
+      <div className="max-w-5xl mx-auto space-y-6">
+        <Breadcrumb>
+          <BreadcrumbList>
+            <BreadcrumbItem>
+              <BreadcrumbLink href="/student">Student</BreadcrumbLink>
+            </BreadcrumbItem>
+            <BreadcrumbSeparator />
+            <BreadcrumbItem>
+              <BreadcrumbPage>Grades</BreadcrumbPage>
+            </BreadcrumbItem>
+          </BreadcrumbList>
+        </Breadcrumb>
+        <h1 className="text-2xl font-semibold">My Results</h1>
+        {loading ? (
+          <div className="space-y-3">
+            <Skeleton className="h-6 w-40" />
+            <Skeleton className="h-40 w-full" />
           </div>
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-            <div className="text-center p-4 bg-blue-50 rounded-lg">
-              <p className="text-sm text-gray-600">Overall Grade</p>
-              <p className="text-2xl font-bold text-blue-600">{calculateOverallGrade()}</p>
-            </div>
-            <div className="text-center p-4 bg-green-50 rounded-lg">
-              <p className="text-sm text-gray-600">Completed</p>
-              <p className="text-2xl font-bold text-green-600">
-                {gradesData.filter(g => g.status === "Graded").length}/{gradesData.length}
-              </p>
-            </div>
-            <div className="text-center p-4 bg-yellow-50 rounded-lg">
-              <p className="text-sm text-gray-600">Pending</p>
-              <p className="text-2xl font-bold text-yellow-600">
-                {gradesData.filter(g => g.status === "Submitted" || g.status === "Not Submitted").length}
-              </p>
-            </div>
-          </div>
-        </div>
-
-        {/* Grades Table */}
-        <div className="rounded-md border border-gray-200 bg-white shadow-sm">
-          <div className="border-b border-gray-200 px-6 py-4">
-            <h2 className="text-lg font-semibold text-gray-900">Assignment Grades</h2>
-          </div>
-          
-          <div className="overflow-x-auto">
-            <table className="w-full">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Assignment
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Score
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Percentage
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Status
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Due Date
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Submitted
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="bg-white divide-y divide-gray-200">
-                {gradesData.map((grade) => (
-                  <tr key={grade._id} className="hover:bg-gray-50">
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <div className="text-sm font-medium text-gray-900">{grade.title || ''}</div>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <span className="text-sm font-medium text-green-600">
-                        {typeof grade.score === 'number' && typeof grade.max_points === 'number' ? `${grade.score}/${grade.max_points}` : ''}
-                      </span>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <span className="text-sm text-green-600">
-                        {typeof grade.score === 'number' && typeof grade.max_points === 'number' ? `${Math.round((grade.score / grade.max_points) * 100)}%` : ''}
-                      </span>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <span className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${
-                        grade.status === "Graded" ? "bg-green-100 text-green-800" :
-                        grade.status === "Submitted" ? "bg-yellow-100 text-yellow-800" :
-                        "bg-red-100 text-red-800"
-                      }`}>
-                        {grade.status}
-                      </span>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                      {grade.due_date ? new Date(grade.due_date).toLocaleDateString() : ''}
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                      {grade.submitted_at ? new Date(grade.submitted_at).toLocaleDateString() : ''}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </main>
+        ) : (
+          <ResultsTable rows={sorted} />
+        )}
+      </div>
     </div>
   )
-} 
+}
